pyproject.toml
[tool.poetry]
name = "ad-classification"
version = "0.1.0"
description = ""
authors = ["Anton Roman Young <89807629+antonryoung02@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.4"
torch = "^2.1.2"
pytest = "^7.4.3"
pillow = "^10.1.0"
torchvision = "^0.16.2"
matplotlib = "^3.8.2"
scikit-learn = "^1.3.2"
seaborn = "^0.13.0"
pylint = "^3.0.3"
tqdm = "^4.66.1"
opencv-python = "^4.8.1.78"
tensorboard = "^2.16.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
